error in consul deletion

+++ internal/topic/core.go
@@ -186,13 +186,16 @@ func (c *Core) DeleteTopic(ctx context.Context, m *Model) error {
 		return err
 	}
 
+	if m.IsRetryTopic() || m.IsSubscriptionInternalTopic() {
+		return nil
+	}
+
 	if ok, err := c.Exists(ctx, m.Key()); !ok {
 		if err != nil {
 			return err
 		}
-		return nil
+		return merror.New(merror.NotFound, "Topic not found")
 	}
-
 	return c.repo.Delete(ctx, m)
 }
